Add hook for ignoring some peers and ssl-errors
Introduce method `isPeerAccepted` to let subclasses of Tufao::WebSocket
allow/disallow peers with specific certificates. This change enables
example client software to implement certificate pinning.

Furthermore, introducing method `shouldIgnoreErrors` enables subclasses
of Tufao::WebSocket to change the semantics of the underyling private
`onSslErrors` method: It can ignore SSL errors, which would not be
ignored by `onSslErrors`, causing a connection to be closed and a
peer to be rejected.